+++ frontend/src/core/registry.js
@@ -18,6 +18,7 @@ const MODULES = {
   'parallax':         () => import('../modules/parallax.js'),
   'split':       () => import('../modules/split-text.js'),
   'scatter':     () => import('../modules/scatter.js'),
+  'scramble':    () => import('../modules/scramble.js'),
   'text-reveal': () => import('../modules/text-reveal.js'),
   'typewriter':  () => import('../modules/typewriter.js'),
   'float':       () => import('../modules/float.js'),